all: apply various fixes from noisy linters
Basically minor cleanups to remove redundancy.
Also stop using the deprecated ioutil package.

+++ multirpc/transports/subpubtransport/subpub.go
@@ -42,7 +42,7 @@ func (p *SubPubHandle) Init(c *transports.Connection) error {
 		p.Conn.Port = 45678
 	}
 	private := p.Conn.Encryption == "private"
-	sp := subpub.NewSubPub(s.Private, []byte(p.Conn.TransportKey), int32(p.Conn.Port), private)
+	sp := subpub.NewSubPub(s.Private, []byte(p.Conn.TransportKey), p.Conn.Port, private)
 	c.Address = sp.PubKey
 	p.SubPub = sp
 	return nil
